requirements.txt
Django==2.1.11
gunicorn==19.9.0
ffmpeg-python
sox
youtube-dl
beautifulsoup4
scipy